Support fallback to old version of zmq
