+++ package.json
@@ -2136,6 +2136,7 @@
         "lodash": "^4.17.21",
         "marked": "^4.0.10",
         "node-fetch": "^2.6.7",
+        "node-gyp-build": "^4.6.0",
         "node-stream-zip": "^1.6.0",
         "pdfkit": "^0.13.0",
         "pidtree": "^0.6.0",
@@ -2178,7 +2179,8 @@
         "vscode-languageserver-protocol": "3.17.2-next.6",
         "vscode-tas-client": "^0.1.27",
         "ws": "^6.2.2",
-        "zeromq": "^6.0.0-beta.16"
+        "zeromq": "^6.0.0-beta.16",
+        "zeromqold": "npm:zeromq@^6.0.0-beta.6"
     },
     "devDependencies": {
         "@actions/core": "^1.9.1",
